feat: update publish endpoint to include post id parameter
The publish endpoint has been modified to include the post id in the route path as a required parameter, changing from "/publish" to "/:id/publish". This allows for more precise targeting of specific posts when updating their published status and follows RESTful API conventions by including the resource identifier in the URL path.